+++ projects/forms.py
@@ -23,28 +23,7 @@ class AddProject(ModelForm):
                 }
             )
 
-        self.fields["url"].widget.attrs.update(
-            {"placeholder": "https://test.com"}
-        )
-
-    def save(self):
-        instance = super(AddProject, self).save()
-
-        @transaction.on_commit
-        def send_notification():
-            message = f"""
-            {instance.user_email} submitted a project.
-            Instance: {instance}
-          """
-            send_mail(
-                "New Project Submission",
-                message,
-                "rasul@builtwithdjango.com",
-                ["rasul@builtwithdjango.com"],
-                fail_silently=False,
-            )
-
-        return instance
+        self.fields["url"].widget.attrs.update({"placeholder": "https://test.com"})
 
     class Meta:
         model = Project
@@ -69,33 +48,6 @@ class AddComment(ModelForm):
                     "rows": "3",
                 }
             )
-
-    def save(self):
-        instance = super(AddComment, self).save()
-
-        @transaction.on_commit
-        def send_notification():
-
-            try:
-                project_owner_email = Project.objects.filter(
-                    title=instance.project
-                )[0].maker.user.email
-            except AttributeError:
-                project_owner_email = "rasul@builtwithdjango.com"
-
-            message = f"""
-            {instance.author} left a comment on your project ({instance.project}).
-            Comment: {instance.comment}
-          """
-            send_mail(
-                "New Comment on your Project",
-                message,
-                "rasul@builtwithdjango.com",
-                ["rasul@builtwithdjango.com", project_owner_email],
-                fail_silently=False,
-            )
-
-        return instance
 
     class Meta:
         model = Comment
@@ -134,9 +86,7 @@ class ProjectUpdateViewForm(ModelForm):
 
         self.fields["is_for_sale"].help_text = None
         self.fields["is_for_sale"].widget.attrs.update(
-            {
-                "class": "focus:ring-indigo-500 h-4 w-4 text-indigo-600 border-gray-300 rounded"
-            }
+            {"class": "focus:ring-indigo-500 h-4 w-4 text-indigo-600 border-gray-300 rounded"}
         )
 
     class Meta: